LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-preflight-star.any.worker-expected.txt

CONSOLE MESSAGE: Method OK is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Method PUT is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Request header field X-Test is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: Request header field X-Test is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: Method PUT is not allowed by Access-Control-Allow-Methods.

PASS CORS that succeeds with credentials: false; method: GET (allowed: get); header: X-Test,1 (allowed: x-test) 
PASS CORS that succeeds with credentials: false; method: SUPER (allowed: *); header: X-Test,1 (allowed: x-test) 
PASS CORS that succeeds with credentials: false; method: OK (allowed: *); header: X-Test,1 (allowed: *) 
PASS CORS that fails with credentials: true; method: OK (allowed: *); header: X-Test,1 (allowed: *) 
PASS CORS that fails with credentials: true; method: PUT (allowed: *); header:  (allowed: ) 
PASS CORS that succeeds with credentials: true; method: PUT (allowed: PUT); header:  (allowed: *) 
PASS CORS that fails with credentials: true; method: GET (allowed: get); header: X-Test,1 (allowed: *) 
PASS CORS that fails with credentials: true; method: GET (allowed: *); header: X-Test,1 (allowed: *) 
PASS CORS that succeeds with credentials: true; method: * (allowed: *); header: *,1 (allowed: *) 
PASS CORS that fails with credentials: true; method: PUT (allowed: put); header:  (allowed: *) 
